+++ hdf5-derive/src/lib.rs
@@ -20,10 +20,19 @@ pub fn derive(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let (impl_generics, ty_generics, where_clause) = input.generics.split_for_impl();
     let body = impl_trait(&name, &input.data, &input.attrs, &ty_generics);
     let dummy = Ident::new(&format!("_IMPL_H5TYPE_FOR_{}", name), Span::call_site());
+
+    // Determine name of parent crate, even if renamed using "package"
+    let crate_name = match proc_macro_crate::crate_name("hdf5-metno").unwrap() {
+        proc_macro_crate::FoundCrate::Itself => quote!(::hdf5_metno),
+        proc_macro_crate::FoundCrate::Name(name) => {
+            let ident = Ident::new(&name, Span::call_site());
+            quote!( ::#ident )
+        }
+    };
     let expanded = quote! {
         #[allow(dead_code, unused_variables, unused_attributes)]
         const #dummy: () = {
-            use ::hdf5_metno as _h5;
+            use #crate_name as _h5;
 
             #[automatically_derived]
             unsafe impl #impl_generics _h5::types::H5Type for #name #ty_generics #where_clause {